Move consider-using-augmented-assign to CodeStyle extension
