+++ doc/data/messages/c/consider-using-augmented-assign/pylintrc
@@ -1,0 +1,3 @@
+[MAIN]
+load-plugins=pylint.extensions.code_style
+enable=consider-using-augmented-assign
