--- doc/whatsnew/fragments/3391.new_check
@@ -1,4 +1,0 @@
-Added ``consider-using-augmented-assign`` which flags ``x = x + 1`` to simplify to
-``x += 1``.
-
-Closes #3391
