+++ tests/functional/ext/code_style/cs_consider_using_augmented_assign.rc
@@ -1,0 +1,3 @@
+[MAIN]
+load-plugins=pylint.extensions.code_style
+enable=consider-using-augmented-assign
