+++ tests/functional/ext/code_style/cs_consider_using_augmented_assign.txt
