+++ tests/functional/ext/code_style/cs_default.py
@@ -1,0 +1,6 @@
+"""Test default configuration for code-style checker."""
+# pylint: disable=invalid-name
+
+# consider-using-augmented-assign is disabled by default
+x = 1
+x = x + 1
