+++ tests/functional/ext/code_style/cs_default.rc
@@ -1,0 +1,2 @@
+[MAIN]
+load-plugins=pylint.extensions.code_style
