+++ tests/functional/u/undefined/undefined_variable_py38.py
@@ -1,5 +1,5 @@
 """Tests for undefined variable with assignment expressions"""
-# pylint: disable=using-constant-test, expression-not-assigned, consider-using-augmented-assign
+# pylint: disable=using-constant-test, expression-not-assigned
 
 # Tests for annotation of variables and potentially undefinition
