+++ tests/functional/u/used/used_before_assignment_conditional.py
@@ -1,7 +1,5 @@
 """used-before-assignment cases involving IF conditions"""
 
-# pylint: disable=consider-using-augmented-assign
-
 if 1 + 1 == 2:
     x = x + 1  # [used-before-assignment]
